Feature/update query tool raw import process/puls 4848 (#208)
* Add validation for key slug1; in seed data excel sheet

* Edit Orion import sanitizeKeysAndTrimData util to not camelCase _id

* Convert _id to ObjectId in upload resolver

* Get bulk upsertion working for CSVs going into pulseRaw.organizationsTest

* Tweak target raw collection name

* Adjust xlsx sheet_to_json config, sanitizeKeysAndTrimData to default empty cells to null while keeping all keys

* Feature/set up auto update of core orgs/puls 4850 (#210)

* Create changeStreams directory and move  WIP logic there

* Get insertion preliminarily working for new connection

* Break out insert operation for readability, add transactiion

* Implement working bilateral update logic

* Revise update operation to take into account dupe provider/payers

* Add console.logs to update operation

* Change sheet_to_json to make defval null ofor empty cells rather than excluding those key-value pairs

* Revise sanitizeKeysAndTrimData util to set falsey values other than false to null

* Replace insert and update files with upsert file; adjust changeStream top-level

* Tweak comments in upsert util code

*  Add temp warning text to import excel sheet card

* Switch to real organizations collection

+++ src/backend/resolvers/mutations/collection/upload.js
@@ -1,6 +1,7 @@
 const _ = require('lodash')
 const UserInputError = require('apollo-server-express').UserInputError
 const getErrorObj = require('../../../validation/getErrorObj')
+const { ObjectId } = require('mongodb')
 
 const uploadCollection = async (
   parent,
@@ -22,6 +23,20 @@ const uploadCollection = async (
   }
 
   const targetCollection = pulseRawDb.collection(collectionName)
+
+  if (collectionName === 'orgConnections') {
+    const bulkOp = targetCollection.initializeOrderedBulkOp();
+
+    data.forEach(({ _id, ...rest }) => {
+      const targetId = _id ? ObjectId(_id) : new ObjectId()
+
+      bulkOp.find({ _id: targetId }).upsert().updateOne({ $set: rest })
+    })
+
+    await bulkOp.execute()
+
+    return
+  }
 
   await targetCollection.deleteMany()
   await targetCollection.insertMany(data)
